Replace DepotDownloader with SteamCMD
Steam recently started requiring downloaders to provide request codes for manifests. This means DepotDownloader crashed when not given a manifest ID. However, this package relies on being able to simply download the latest manifest for a given beta branch, and not have to explicitly provide it. Since DepotDownloader doesn't seem to be very active anymore, it is replaced with the official tool SteamCMD.